REALLY BIG PUSH!
NOTE: Creating a bunch of different functions is easier for me to understand. I can differentiate between different components of the game this way.

LOTS and LOTS of research *sigh.

-translating tetrimino lists created above to actual game pieces

-creating detection for checking if the piece is "lost" (aka not within the bounds of the playing grid)

-detecting if the current piece can be moved; checks if the tetrimino (full) is on the grid. IF not, the piece will not move!

-creating a fall time for the block so it can move down the board. Also stops the piece where it hits another color rather than black (indicates tetrimino is there) and locks it into place.

-codes in for quitting if you lose the game.

-gives a startup page before the game!

-a bit of debugging, adding in things I missed before + renaming things to make it more clear.

MIGHT be forgetting some stuff. But the code still is crashing :( theoretically everything seems okay. I'll do a big debugging thing @ the end of coding all the dif. functions